--- a/src/Clp.jl
+++ b/src/Clp.jl
@@ -9,6 +9,7 @@ using Compat
 # This 'using' is required to suppress a warning about Clp not having Libdl in its
 # dependencies (Libdl is used by BinaryProvider), e.g.: bicycle1885/CodecZlib.jl#26.
 using Libdl
+import SparseArrays
 
 include("ClpCInterface.jl")
 include("ClpSolverInterface.jl")