--- a/test/MOIWrapper.jl
+++ b/test/MOIWrapper.jl
@@ -26,12 +26,16 @@ end
         "linear1",
         # linear10 test is tested below because it has interval sets.
         "linear10",
+        "linear10b",
         # linear11 test is excluded as it fails on Linux for some reason.
         # It passes on Mac and Windows.
         "linear11",
         # linear12 test requires the InfeasibilityCertificate for variable
         # bounds. These are available through C++, but not via the C interface.
-        "linear12"
+        "linear12",
+        # partial_start requires VariablePrimalStart to be implemented by the
+        # solver.
+        "partial_start"
     ])
 
     @testset "Interval Bridge" begin
@@ -49,7 +53,7 @@ end
     end
     @testset "emptytest" begin
         MOIT.emptytest(solver)
-    end 
+    end
     @testset "copytest" begin
         solver2 = Clp.Optimizer(LogLevel = 0)
         MOIT.copytest(solver,solver2)