This significantly improves the tine it takes to copy the model when in caching mode.   Mostly this is achieved through the use of add_rows & add_columns rather than their singular counterparts. 

This pull is dependant on the following PR:
https://github.com/JuliaOpt/LinQuadOptInterface.jl/pull/96
https://github.com/JuliaOpt/MathOptInterface.jl/pull/696

Benchmarking will follow.