@ndinsmore 2019-03-12
The benchmark bellow shows that there is about a 40x speed increase in the time it takes to copy the model from the caching bridge to clp.

Benchmark Code:
```julia
using JuMP
using Clp
using GLPK
using BenchmarkTools
using Profile
using Random
using SparseArrays

function  build_model(solver, num_variables, num_constraints, density; opt_options::NamedTuple=NamedTuple())
    model = Model(with_optimizer(solver.Optimizer; opt_options...))
    # model = Model(with_optimizer(solver.Optimizer; opt_options...))
    rng=Random.MersenneTwister(1)

    sp_mat=sprand(rng,num_constraints,num_variables,density)
    # display(Matrix(sp_mat))
    vars=@variable(model,0<=vars[i=1:num_variables]<=0)

    cons=@constraint(model,cons,sp_mat * vars .== 0)
    return model
end
#-------
num_vars=10000
num_cons=10000
density=.001
@benchmark optimize!(model) setup=(model=build_model(Clp,$num_vars,$num_cons,$density,opt_options=(LogLevel=0,MaximumSeconds=10.0)))
```

With master on MOI, LQOI, & Clp:
```julia
julia> @benchmark optimize!(model) setup=(model=build_model(Clp,$num_vars,$num_cons,$density,opt_options=(LogLevel=0,MaximumSeconds=10.0)))
BenchmarkTools.Trial: 
  memory estimate:  3.05 GiB
  allocs estimate:  1350058
  --------------
  minimum time:     3.940 s (13.25% GC)
  median time:      3.959 s (12.61% GC)
  mean time:        3.959 s (12.61% GC)
  maximum time:     3.978 s (11.97% GC)
  --------------
  samples:          2
  evals/sample:     1
```

Benchmark after applying the above PRs:
```julia
julia> @benchmark optimize!(model) setup=(model=build_model(Clp,$num_vars,$num_cons,$density,opt_options=(LogLevel=0,MaximumSeconds=10.0)))
BenchmarkTools.Trial: 
  memory estimate:  32.56 MiB
  allocs estimate:  780109
  --------------
  minimum time:     130.490 ms (10.37% GC)
  median time:      150.618 ms (15.79% GC)
  mean time:        154.012 ms (17.83% GC)
  maximum time:     187.498 ms (25.74% GC)
  --------------
  samples:          16
  evals/sample:     1
```