@ndinsmore 2019-04-09
Benchmarking code:
```julia
using JuMP
using Clp
using GLPK
using BenchmarkTools
using Random
using SparseArrays
using Profile

function  build_model(solver, sp_mat; opt_options::NamedTuple=NamedTuple())
    model = Model(with_optimizer(solver.Optimizer; opt_options...))
    # model = Model(with_optimizer(solver.Optimizer; opt_options...))
    num_variables=size(sp_mat,2)
    # display(Matrix(sp_mat))
    vars=@variable(model,0<=vars[i=1:num_variables]<=0)

    cons=@constraint(model,cons,sp_mat * vars .== 0)
    return model
end
function  build_model_direct(solver, sp_mat; opt_options::NamedTuple=NamedTuple())
    model = JuMP.direct_model(solver.Optimizer( ;opt_options...))
    # model = Model(with_optimizer(solver.Optimizer; opt_options...))
    num_variables=size(sp_mat,2)
    # display(Matrix(sp_mat))
    vars=@variable(model,0<=vars[i=1:num_variables]<=0)

    cons=@constraint(model,cons,sp_mat * vars .== 0)
    return model
end
num_vars=10000
num_cons=10000
density=.001
rng=Random.MersenneTwister(1)
sp_mat=sprand(rng,num_cons,num_vars,density)

direct_build_clp = @benchmark model=build_model_direct(Clp,sp_mat,opt_options=(LogLevel=1,MaximumSeconds=10.0))
build_clp = @benchmark model=build_model(Clp,$sp_mat,opt_options=(LogLevel=1,MaximumSeconds=10.0))
opt_clp =@benchmark optimize!(model) setup=(model=build_model(Clp,$sp_mat,opt_options=(LogLevel=0,MaximumSeconds=10.0)))
```
Master:
```julia
julia> direct_build_clp = @benchmark model=build_model_direct(Clp,sp_mat,opt_options=(LogLevel=1,MaximumSeconds=10.0))
BenchmarkTools.Trial: 
  memory estimate:  1.56 GiB
  allocs estimate:  959800
  --------------
  minimum time:     2.868 s (6.70% GC)
  median time:      3.095 s (7.84% GC)
  mean time:        3.095 s (7.84% GC)
  maximum time:     3.323 s (8.83% GC)
  --------------
  samples:          2
  evals/sample:     1

julia> build_clp = @benchmark model=build_model(Clp,$sp_mat,opt_options=(LogLevel=1,MaximumSeconds=10.0))
BenchmarkTools.Trial: 
  memory estimate:  30.32 MiB
  allocs estimate:  407369
  --------------
  minimum time:     50.563 ms (10.68% GC)
  median time:      68.254 ms (15.08% GC)
  mean time:        71.712 ms (16.95% GC)
  maximum time:     187.390 ms (57.66% GC)
  --------------
  samples:          70
  evals/sample:     1

julia> opt_clp =@benchmark optimize!(model) setup=(model=build_model(Clp,$sp_mat,opt_options=(LogLevel=0,MaximumSeconds=10.0)))
BenchmarkTools.Trial: 
  memory estimate:  43.94 MiB
  allocs estimate:  900321
  --------------
  minimum time:     2.318 s (1.52% GC)
  median time:      2.323 s (1.85% GC)
  mean time:        2.354 s (2.92% GC)
  maximum time:     2.421 s (5.29% GC)
  --------------
  samples:          3
  evals/sample:     1
```
With above PRs:
```julia 
julia> direct_build_clp = @benchmark model=build_model_direct(Clp,sp_mat,opt_options=(LogLevel=1,MaximumSeconds=10.0))
BenchmarkTools.Trial: 
  memory estimate:  1.56 GiB
  allocs estimate:  979800
  --------------
  minimum time:     3.035 s (9.45% GC)
  median time:      3.035 s (8.12% GC)
  mean time:        3.035 s (8.12% GC)
  maximum time:     3.036 s (6.79% GC)
  --------------
  samples:          2
  evals/sample:     1

julia> build_clp = @benchmark model=build_model(Clp,$sp_mat,opt_options=(LogLevel=1,MaximumSeconds=10.0))
BenchmarkTools.Trial: 
  memory estimate:  30.32 MiB
  allocs estimate:  407369
  --------------
  minimum time:     53.355 ms (12.94% GC)
  median time:      67.650 ms (15.54% GC)
  mean time:        69.863 ms (17.19% GC)
  maximum time:     160.476 ms (60.86% GC)
  --------------
  samples:          72
  evals/sample:     1

julia> opt_clp =@benchmark optimize!(model) setup=(model=build_model(Clp,$sp_mat,opt_options=(LogLevel=0,MaximumSeconds=10.0)))
BenchmarkTools.Trial: 
  memory estimate:  32.56 MiB
  allocs estimate:  780344
  --------------
  minimum time:     116.185 ms (15.59% GC)
  median time:      163.528 ms (22.02% GC)
  mean time:        170.163 ms (26.51% GC)
  maximum time:     284.184 ms (54.23% GC)
  --------------
  samples:          18
  evals/sample:     1
```